tpm2

Trait UnmarshalFixed

pub trait UnmarshalFixed: MarshalFixed + for<'src> Unmarshal<'src> {
    // Required method
    fn unmarshal_fixed(
        &mut self,
        arr: &Self::Array,
    ) -> Result<(), UnmarshalError>;

    // Provided method
    fn unmarshal_value(arr: &Self::Array) -> Result<Self, UnmarshalError>
       where Self: Default { ... }
}

Similar to MarshalFixed but for unmarshalling

This trait doesn’t have a 'src lifetime parameter like Unmarshal, as types with fixed SIZE don’t need to retain references to the source buffer.

Required Methods

fn unmarshal_fixed(&mut self, arr: &Self::Array) -> Result<(), UnmarshalError>

This can still fail if the source buffer has a bad value, but we know we have a buffer of the correct length.

Provided Methods

fn unmarshal_value(arr: &Self::Array) -> Result<Self, UnmarshalError>
where Self: Default,

Helper function (nicer now that we don’t need to return the unused buffer).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.